+++ src/ppocr/box_char/box.hpp
@@ -39,16 +39,16 @@ struct Box
     : bounds_(bnd)
     {}
 
-    size_t x() const noexcept { return idx_.x(); }
-    size_t y() const noexcept { return idx_.y(); }
+    unsigned x() const noexcept { return idx_.x(); }
+    unsigned y() const noexcept { return idx_.y(); }
 
-    size_t w() const noexcept { return bounds_.w(); }
-    size_t h() const noexcept { return bounds_.h(); }
+    unsigned w() const noexcept { return bounds_.w(); }
+    unsigned h() const noexcept { return bounds_.h(); }
 
-    size_t bottom() const noexcept { return y()+h()-1; }
-    size_t right() const noexcept { return x()+w()-1; }
-    size_t left() const noexcept { return x(); }
-    size_t top() const noexcept { return y(); }
+    unsigned bottom() const noexcept { return y()+h()-1; }
+    unsigned right() const noexcept { return x()+w()-1; }
+    unsigned left() const noexcept { return x(); }
+    unsigned top() const noexcept { return y(); }
 
     Index const & index() const noexcept { return idx_; }
     Bounds const & bounds() const noexcept { return bounds_; }
